requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pymongo
bcrypt